+++ database/migrations/2020_11_05_232739_create_q3w_material_operations_table.php
@@ -116,7 +116,7 @@ class CreateQ3wMaterialOperationsTable extends Migration
         Schema::create('q3w_operation_materials', function (Blueprint $table) {
             $table->bigIncrements('id')->comment('Уникальный идентификатор');
             $table->bigInteger('material_operation_id')->unsigned()->comment('Идентификатор операции')->index();
-            $table->bigInteger('standard_id')->unsigned()->comment('Идентификатор эталона')->index();
+            $table->integer('standard_id')->unsigned()->comment('Идентификатор эталона')->index();
             $table->integer('amount')->unsigned()->nullable()->comment('Количество в штуках (для штучного учета)');
             $table->double('quantity')->unsigned()->comment('Количество в единицах измерения');
